Refactored Alert Dialog, now Alert Modal and simplified API
