--- src/components/AlertDialog/index.ts
@@ -1,2 +1,0 @@
-export * from "./AlertDialog";
-export { default } from "./AlertDialog";
