+++ src/components/AlertModal/index.ts
@@ -1,0 +1,2 @@
+export * from "./AlertModal";
+export { default } from "./AlertModal";
